moved AudioKit Start/Stop functions to their own extension file
