+++ AudioKit/Common/Internals/AudioKit.swift
@@ -41,7 +41,7 @@ public typealias AKCallback = () -> Void
     @objc static var finalMixer = AKMixer()
 
     /// Notification observers
-    fileprivate static var notificationObservers: [Any] = []
+    internal static var notificationObservers: [Any] = []
 
     /// An audio output operation that most applications will need to use last
     @objc open static var output: AKNode? {
@@ -211,159 +211,6 @@ public typealias AKCallback = () -> Void
         #else
             //not available on ios
         #endif
-    }
-
-    // MARK: - Start/Stop
-
-    /// Start up the audio engine with periodic functions
-    open static func start(withPeriodicFunctions functions: AKPeriodicFunction...) throws {
-        for function in functions {
-            function.connect(to: finalMixer)
-        }
-        try start()
-    }
-
-    /// Start up the audio engine
-    @objc open static func start() throws {
-        if output == nil {
-            AKLog("No output node has been set yet, no processing will happen.")
-        }
-        // Start the engine.
-        try AKTry {
-            engine.prepare()
-        }
-
-        #if os(iOS)
-
-            if AKSettings.enableRouteChangeHandling {
-
-                let routeChangeObserver = NotificationCenter.default.addObserver(forName: .AVAudioSessionRouteChange,
-                                                                                 object: nil,
-                                                                                 queue: OperationQueue.main,
-                                                                                 using: { (notification) in
-                                                                                    AudioKit.restartEngineAfterRouteChange(notification)
-                })
-                notificationObservers.append(routeChangeObserver)
-
-            }
-
-            if AKSettings.enableCategoryChangeHandling {
-                let configurationChangeObserver = NotificationCenter.default.addObserver(forName: .AVAudioEngineConfigurationChange,
-                                                                                         object: engine,
-                                                                                         queue: OperationQueue.main,
-                                                                                         using: { (notification) in
-                                                                                            AudioKit.restartEngineAfterConfigurationChange(notification)
-                })
-                notificationObservers.append(configurationChangeObserver)
-            }
-            try updateSessionCategoryAndOptions()
-            try AVAudioSession.sharedInstance().setActive(true)
-        #endif
-
-        try AKTry {
-            try engine.start()
-        }
-        shouldBeRunning = true
-    }
-
-    @objc fileprivate static func updateSessionCategoryAndOptions() throws {
-        #if !os(macOS)
-            let sessionCategory = AKSettings.computedSessionCategory()
-
-            #if os(iOS)
-                let sessionOptions = AKSettings.computedSessionOptions()
-                try AKSettings.setSession(category: sessionCategory,
-                                          with: sessionOptions)
-            #elseif os(tvOS)
-                try AKSettings.setSession(category: sessionCategory)
-            #endif
-        #endif
-    }
-
-    /// Stop the audio engine
-    @objc open static func stop() throws {
-        // Stop the engine.
-        try AKTry {
-            engine.stop()
-        }
-        shouldBeRunning = false
-
-        #if os(iOS)
-        notificationObservers.forEach { (observer) in
-            NotificationCenter.default.removeObserver(observer)
-        }
-
-        do {
-            try AVAudioSession.sharedInstance().setActive(false)
-        } catch {
-            AKLog("couldn't stop session \(error)")
-            throw error
-        }
-        #endif
-    }
-
-
-    // MARK: - Configuration Change Response
-
-    // Listen to changes in audio configuration
-    // and restart the audio engine if it stops and should be playing
-    @objc fileprivate static func restartEngineAfterConfigurationChange(_ notification: Notification) {
-        do {
-            if shouldBeRunning && !engine.isRunning {
-                #if !os(macOS)
-                    let appIsNotActive = UIApplication.shared.applicationState != .active
-                    let appDoesNotSupportBackgroundAudio = !AKSettings.appSupportsBackgroundAudio
-
-                    if appIsNotActive && appDoesNotSupportBackgroundAudio {
-                        AKLog("engine not restarted after configuration change since app was not active and does not support background audio")
-                        return
-                    }
-                #endif
-
-                try engine.start()
-
-                // Sends notification after restarting the engine, so it is safe to resume AudioKit functions.
-                if AKSettings.notificationsEnabled {
-                    NotificationCenter.default.post(
-                        name: .AKEngineRestartedAfterConfigurationChange,
-                        object: nil,
-                        userInfo: notification.userInfo)
-                }
-            }
-        } catch {
-            AKLog("error restarting engine after route change")
-            // Note: doesn't throw since this is called from a notification observer
-        }
-    }
-
-    // Restarts the engine after audio output has been changed, like headphones plugged in.
-    @objc fileprivate static func restartEngineAfterRouteChange(_ notification: Notification) {
-        if shouldBeRunning && !engine.isRunning {
-            do {
-                #if !os(macOS)
-                    let appIsNotActive = UIApplication.shared.applicationState != .active
-                    let appDoesNotSupportBackgroundAudio = !AKSettings.appSupportsBackgroundAudio
-
-                    if appIsNotActive && appDoesNotSupportBackgroundAudio {
-                        AKLog("engine not restarted after route change since app was not active and does not support background audio")
-                        return
-                    }
-                #endif
-
-                try engine.start()
-
-                // Sends notification after restarting the engine, so it is safe to resume AudioKit functions.
-                if AKSettings.notificationsEnabled {
-                    NotificationCenter.default.post(
-                        name: .AKEngineRestartedAfterRouteChange,
-                        object: nil,
-                        userInfo: notification.userInfo)
-                }
-            } catch {
-                AKLog("error restarting engine after route change")
-                // Note: doesn't throw since this is called from a notification observer
-            }
-        }
     }
 
     // MARK: - Disconnect node inputs
